+++ test/unit_test/engine/test_engine.py
@@ -208,7 +208,7 @@ class TestEngine:
             # ensure two converted models are from the same input model
             assert actual_res.nodes[model_id].parent_model_id == input_model_id
 
-            assert engine.get_model_json_path(actual_res.nodes[model_id].model_id).exists()
+            assert engine.cache.get_model_json_path(actual_res.nodes[model_id].model_id).exists()
             for k, v in result.items():
                 if k == "metrics":
                     assert actual_res.nodes[model_id].metrics.if_goals_met
@@ -471,7 +471,7 @@ class TestEngine:
 
         engine.initialize()
 
-        assert engine._new_model_number == 101
+        assert engine.cache.new_model_number == 101
         assert mock_unlink.call_count == 1
         assert mock_glob.call_count == 2
 
@@ -486,13 +486,11 @@ class TestEngine:
             "clean_evaluation_cache": True,
             "evaluator": evaluator_config,
         }
-        engine = Engine(**options)
-        engine.register(OnnxConversion, clean_run_cache=True)
         with patch.object(Path, "glob"):
             Path.glob.return_value = [Path("cache") / "output" / "435d_0.json"]
 
             with pytest.raises(ValueError) as exc_info:  # noqa: PT011
-                engine.initialize()
+                Engine(**options)
             assert str(exc_info.value) == "invalid literal for int() with base 10: '435d'"
 
     @patch("olive.systems.local.LocalSystem")
